Enhance license moved dependent license to different py file
